refactor: reorganize source tree into core/services/ui/app layers
Move files into layered structure:
- core/{sysinfo,settings,logging}/ — pure logic, no QtWidgets
- services/                        — non-UI lifecycle services
- ui/                              — QtWidgets-dependent code
- app/                             — composition root (unchanged)

Renames:
- utils/utils.{h,cpp}                       -> core/sysinfo/system_info.{h,cpp}
- settings/                                 -> core/settings/
- logger/                                   -> core/logging/
- integration_server/integration_server.*   -> services/integration_server.*
- tray/                                     -> ui/
- tests/tst_utils.cpp                       -> tests/tst_system_info.cpp

Update all #include paths and CMakeLists targets accordingly.
The Utils namespace itself is left unchanged for now.
