+++ SysInfo/core/logging/logger.cpp
