+++ SysInfo/core/settings/settings_manager.h
